feat: extract utils
